pyproject.toml
[project]
name = "discord-bot-python"
version = "0.1.0"
description = "Bot de discord"
authors = [
    { name = "Guillermo Merino", email = "guillermomerino2107@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "black>=25.1.0",
    "schedule>=1.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0"
]

[tool.black]
line-length = 100
target-version = ['py311']
skip-string-normalization = true
